pp/: Support for defined operator ..
.. this is complicated because you don't want to expand "defined" as a macro,
or its argument, when it appears in a #if or #elif line directive, whereas you
want to ignore it (and process it regularly) in all other cases. Currently we
implement this by selectively processing PPME_DEFINED tokens which, if present,
cause the macro-expander to ignore the sequence (thus preserving the defined
operator and its argument), and in most other cases, translating the
PPME_DEFINED to a PPME_IDENT (which itself may be translated further to a
PPME_FUNCTION or PPME_OBJECT macro expansion); which will process the defined
directive for macro expansion.
One edge case we'll fix shortly is "#if defined(defined)"