renderbmp.c: tri5 experiment, z-buffer interpolation (but not testing)...
... this implements the z-buffer value integer interpolation over the triangle.
It is used as the color of the pixel after HSV-to-RGB conversion, for display
purposes. Theory and practice seem to align, more thought needs to be given to
the bits of accuracy that are available. Inner-loop is nice.

tri5 is implemented using tri2 as a starting point. The tri4
quadruple-fragments-rasterizer will need to be merged with this once we settle
on the final approach.