renderbmp.c: tri6 experiment - finding the required 97 bits of precision ... (WIP)
... obviously getting hairy, important to note that, while all that "manual"
math is going to be slow'ish, it is slow'ish per triangle-setup, not slow'ish
per fragment, so we should be good'ish.

tri6 is not yet done, goal is to get both the 64-bit implementation and the
full-precision 128 bit implementation running alongside eachother, and to then
find a few test vectors to look for consistency, then drop the 64 bits and look
for precision edge cases (e.g. dial it up to 30 bits, see if it blows.)