renderbmp.c: tri7 experiment & fix...
tri7 has the option to sample at exact pixels the z-buf value and z-buf value
numerator; these have to be exact matches or our math is off.

Also ran a test with a "large" triangle:
(103,3) - (1073741823, 1073741823) - (100, 1073741823).

this is right up to the 30 bit limit we're looking to support. Initial results
revealed a bug which has been fixed, results now appear to be correct for both
z-buffer values and rasterized result.

The bug occurred because we did not cast the x/y coordinates to 64 bits prior
to multiplying them, causing overflows much sooner than intended. Fixed in tri7.