fix: xy direction flip off by one & experiment tri7 refinement...
... switching from x to y, from ascending to descending (or vice-versa) used the
formula s'=b-s-2 , loosely: new numerator = denominator - old numerator - 2. That
-2 appears to have been a reasoning mistake, and should just be - 1.

Suppose we're at 1/4'ths, and would like to change direction. Steps from 1/4 are
0/4, followed by the switch (so two steps). Steps from 1/4 in the opposite direction
is 2/4, 3/4, followed by the switch (so three steps). In other words, 1/4 should map
not onto s'=4-1-2=1, but s'=4-1-1=2, or 2/4ths, when flipping direction.

To test this, tri7 has been adjusted for a *very* gradual slope vertically.
Point 0,0 should have a z of 1. Point 0,max_y should have a z of 0, and to the
right the point 100,0 should also have a z of 1. Consequently, (0,0) should be 1,
(1,0) should be 1, but (0,1) has z going down an infinitesimally small gradient, and
because we still truncate z-buffer values rather than round to nearest, (0,1), should
truncate to 0. Same for (1,1).