rasterizer.c/h: build out triangle rasterizer stage ...
... idea here is to fill the fragment buffer with fragments and, when the
fragment buffer is full, have it be processed as one set. Processing the
fragments in parallel such that overhead cost are amortized over enough
individual fragments, while at the same time keeping the total set of
fragments small enough such that it might fit in L2 or even L1.

This requires that the rasterizer be able to stop when the fragment buffer
is full and resume when it is empty again. This yielding approach for the
rasterizer is implemented using Duff's device, to load context on entry to
the function and evict context as we need to.