clipping_stage.c/h: re-order "known" varyings and add 3 special...
... special ones are sx, sy and sz, which represent the screenspace integer
coordinates of the vertices. These are not affected (or even processed (or
even touched)) by clipping, however, we need to reserve one of each for
every vertex output from the clipping stage. Consequently, we put them here
for now, may change that up later; but this fits nicely for what I'm
planning with viewport transformation.