primitive_assembly.c: fix for restrict ptr type ...
we had a pointer to a restrict pointer to a uint8, but what we want is a
restrict pointer to a restrict pointer to a uint8 - it is not just that the
rgba data pointed to is only accessed through each fragment's individual
pointer, it is also that those individual pointers are only accessed through
the pointer stepping through the fragment's pointers.

Really trying to liberate the compiler into optimizing this area, whether it
has the desired effect will be seen.