Stencil and depth tests (WIP) ...
... this code is more verbose than I'd like due to wanting to support 16, 24
and 32 bit z-buffer depths; these checks create a lot of similar looking code
as I don't want to rely on the compiler to do the loop unswitching as any
failure to do so would be disastrous for performance on optimized builds.

The stencil buffer is currently always set to a fixed 16 bits.

Tests are performed, z-buffer is updated selectively depending on pass or
fail of stencil and the z-buffer function, however, the RGB and stencil are
not yet updated correctly (that is, selectively.)