struct typename ambiguous grammar implementation ...
... this is roughly analogous to the typename ambiguity in C; we look ahead if
a TYPENAME token is accepted, if so, we check if we have an identifier, and if
that identifier resolves to a struct in the symbol table. If it resolves to a
struct, we convert the token into a TYPENAME token and pass it in. Otherwise,
it stays an identifier and the grammar continues as usual.