refactor pp diagnostics to diags.c/h ...
Refactor the diagnostics from the preprocessor to use a separate diagnostics
object. The idea here is two fold, on the one hand, this allows diags to be
shared across different stages (and possibly compilers) in the library, and
on the other, it allows potentially shared stages access to the diagnostics
without having to know about the compiler it is called from.