Apply sl_expr_validate() for constant-expression ...
... thus checking if it is indeed a constant-expression and issuing an error
otherwise. Note that we could also evaluate it here if validation is successful,
not sure if that's what we want to do (would require passing up the evaluated
value for "constant-expression" rather than an expression) - the alternative is
replicating that code at every site where we use a constant-expression, but still
passing up an "sl_expr *" which is less of a surprise.