Changing things around for stmt::prep_ , is now expr_ ...
... originally envisaged prep_ to be the initializer (e.g. the assignment
expression) for any declaration in a "for (int x=123;;)" - this creates
issues because, according to the spec's grammar, that "for-init-statement"
is a "declaration", i.e., it can have multiple declarators with multiple
initializers. In other words, this: "for (int x=123, y=456;;)" is also
valid. Consequently, a single assignment expression wouldn't do, as
"declaration" may return zero-to-many statements.

One option is to convert those on the fly from statements to a large
expression of exop_sequence's, but that would presume a lot about the
structure of the statements returned.

The other option is to turn prep_ into a series of statements, and convert
any other uses of it to expr_, a single expression; which is what we're,
part of the way, doing here.