start-scope error recovery and the parsing for statement ....
Start-scope can be popped from the stack during error recovery, in which case
the productions containing it will not be reduced. That means we still need to
"pop" the scope if it is in the current stack of scopes. That gets messy because
we are more restricted during %destructor's in that we don't have access to
%params like, in this case, the glsl_es1_compiler cc that houses the current
scope.

Parsing the for statement is a bit of a mess because of the many nooks and cranny's
that a for statement can take in GLSL. E.g. "for (int x=3, y=4; int q = x++; ++y) { .. }"
is valid. Thus, unlike, say, C, there are multiple places where the initializers
bubble up statements into their parent for statement. Additionally (like C), many
of the parts of a for statement are optional, adding to the complexity.