sl_execution.c/h: execution context (WIP) ...
... sl_execution will perform the actual execution of a shader (fragment or
vertex or whatnot) - the idea being we execute rows collectively for each
statement and expression node; thus keeping the ALUs fed and minimizing the
overhead of running directly from the syntax tree.

... also in this commit is an experiment for how we'll let the compiler auto-
vectorize our operations for us; sl_binop_snippet_inc.h is a helper for this,
the idea is we analyze the sequence we're trying to run, and, if the next 8
rows are on an 8-row alignment[*], and they are consecutive (e.g. the
execution chain they form is one consecutive block of rows), then run a little
"for" loop to execute the operation in an extremely naïve way as to elicit a
vectorization from the compiler. Currently tested on MSVC 22 to produce the
desired result for simple operations (e.g. it is indeed vectorizing to SSE and
AVX.)

All this still needs to crystallize a fair bit, work in progress.