sl_exec_prep() and sl_exec_join_chains() and register fields ...
... the register fields in float_regs, int_regs_, bool_regs and so on, are for
storing the pointer to the start of the column of data, such that each column
consist of a single float/int/bool/... holding the data per row (so "struct of
arrays" approach rather than "array of structs"); sl_exec_prep() will allocate
the arrays containing the pointers; but expects caller to initialize the
pointers to the true data columns (e.g. in fragment buffer or primitive
assembly or wherever..)

... sl_exec_join_chains() takes two lists of rows, identified by their
indices, and merges them along chain defined by their exec_chain register/
column. In the common case where every row of execution takes the same
branches, this will not execute and exit early (as the execution flow
does not "merge" if everyone takes the same path.)