prepare for lvalues ...
... key challenge here is that for:
"a[x++] += b"
we'd like to hold a lock on the register containing the 'x++' expression so
we compute it only once. While in a C compiler we might model lvalues as
indirections on a pointer (e.g. a[x++] evaluates to a pointer value with an
indirection node on top, which might be used in a DAG style fashion), in this
case there are no pointers, only registers, and the AST more directly
represents what is in the source, without any additional translations.
Consequently, when evaluating "+=", the a[x++] is read, the "x++" value is
remembered, and when assigning the result from "a[x++] + b" we assign it
back into the "x++" remembered value (without evaluating the post-inc again.)