range allocator (WIP) ...
... allocates a range rather than just a value in isolation, like malloc.
Reason we need it is because sl_reg_alloc currently has registers for each
isolated value; so for an array, each element of the array is individually
allocated. The problem is this makes array elements not adhere to any
predictable order in terms of the registers used, consequently, to evaluate
an array indirection (e.g. exop_array_subscript), we need to pass along the
sl_reg_alloc to find the register of an array's element.
Even though GLSL v1 doesn't support it, if, in the future, we wish to support
multiple nested arrays, this story becomes even more complicated (e.g.
indirections of indirections of indirections, and so on) - this is not only
extremely slow, but very confusing.
A range allocator can ask "give me 50 consecutive float registers" and then
use this to lay out 5 fields of 10 floats each, or whatever. The floats
returned are consecutive because they are a range (much like malloc(50) does
not hand out 50 isolated bytes but a consecutive range.) If the evaluator
for exop_array_subscript knows the registers are in a consecutive range, then
it can just do some math to find the right register based on the index
evaluated, instead of one or more layers of indirection.

Work in progress. I find implementing an allocator like this a complex piece
of kit. Current goal is to keep it simple and stay clear of the rocket science
of a c-lib allocator. To that end, we use two red-black trees, one to house the
sizes of the free ranges, the other to house the location of the free ranges.
These two can then be used to free ranges of registers when given the register
range itself, and to allocate any number of registers and fit it into the
smallest available (best fit) space possible, and the lowest if more spaces are
available. Work in progress.