Revamp of sl_reg_alloc to use new ral_range_allocator approach (WIP) ...
... idea here is we accumulate the arrays subscripted to reach a final field
and form one flat array for each final field. So, e.g., suppose we have

struct a {
  struct b b_[3];
};

struct b {
  vec2 x_;
  bool c_;
};

struct a a_[4];

then a_ will result in a layout of a_::b_::x_.x of 4 * 3 consecutive float registers,
and another layout of a_::b_::x_.y of another 4 * 3 consecutive float registers, and
a layout of a_::b_::c_ of 3 * 4 consecutive booleans.

The "consecutiveness" of this layout is important as it simplifies array indexing.