ref_range_allocator.c/h: introduce watermark ..
.. the idea here is for ref_range_allocator to maintain a watermark so that we
know what the highest range was that was allocated. This will help when we need
to allocate the full register file (so we know how many registers to allocate.)
This is different than the watermark_ field in the previous range_allocator.c/h
as that watermark_ field identifier the *current* highest value, and this
watermark_ identifier the highest value ever reached.