Give each reg_alloc_ a local_frame_ flag ..
.. this will help us distinguish between global variables (which are at offset
0 of the register file) and local variables or temps (which are at the offset
of their function invocation's still-to-be-defined register base offset (e.g.
like a stack frame / base pointer.))

.. this is not ideal, in the ideal case we would not need to check this flag,
but instead inline based on the callgraph; that's a processing step further
removed then one I want to make right now (need to zip this up and test &
debug it.) As a happy accident, it means we'll temporarily support recursion in
the shaders.