first steps toward call graph analysis ..
.. this is a bit of a reversion of the happy accident in the previous commit
msg in that actually doing the call graph analysis (and determining up-front
how many registers of each type we'll need) is very helpful in simplifying
other parts. Notably, it allows us to push the setup and allocation of all
registers to a single initialization performed by the caller, and so
externalized the memory management away from the sl_execution. This helps in
our use-cases like the primitive_assembly and fragment_buffer, where the
sl_execution is not intended to directly meddle in the allocation of columns.
So, no happy accident, just moving forward.