Load into rvalues where applicable prior to evaluation ...
... we'll probably rethink this in a bit; particularly, the problem is that by
placing the rvalue_ inside of the reg_alloc, we've created a situation where
a struct might have some members with an rvalue_ and other members with an
offset and no rvalue. While we could write code that handles that in a
tolerant fashion, I'm not sure if that would occur in practice or have any
value. The alternative is for the sl_expr to have an rvalue_, thus an rvalue
can only be conceptualized at the "value of a (sub)expression" level.