various small reorganizations in sl_execution ..
.. sl_execution now takes a pointer to the compilation unit so we can perform
symtable queries using the exec pointer alone. sl_exec_prep does not take the
function to execute, instead, this (and the chain to execute) are now in
sl_exec_run() -- reason being we'd like to execute sl_exec_prep() once after
compilation and call sl_exec_run() whenever we are about to execute. The
initialization based on the function to execute should therefore be a runtime
thing that's separate from all the allocations we need.

sl_exec_run() now also performs the global initialization by their literal
values, and pushes the initial function execution frame, as well as the the
bootstrap execution point.