struct constructor & typename fixes ..
.. cc->is_typename_permitted_ is a flag that maintains whether or not we allow
for the next identifier to be a typename. If the flag is true, and the next
token is an identifier, and the current position in the grammar parsing allows
for a typename, then, prior to parsing, the identifier will be looked up in
the symbol table and, if an associated typename (e.g. struct) is found for the
identifier, the token is converted into a typename token. As soon as that
happens, cc->is_typename_permitted_ is set to false; so subsequent identifiers
cannot automatically convert into a typename.

This solves an ambiguity:
struct blah { .. };
blah blah; // declares an entry "blah" of type "blah"

It is however the grammar's responsibility to reset the
cc->is_typename_permitted_ flag back to true to allow typenames to be used
again. We did not consistently do this.

A similar approach exists for parsing the famous C language typename ambiguity,
hence its naming here.