Keep uniforms in sl_program and export both ...
sl_program_load_uniform_for_execution and
sl_program_load_uniforms_for_execution. Note that loading a uniform requires
access to possibly both shaders, as it requires access to their corresponding
sl_execution to load the variables in. This is why we have this method in
sl_program and not in sl_uniforms (the latter might have seemed a more
intuitive place.)