glCopyTexImage2D() ..
.. also, add all faces of the cube map as sampler_2d's to the gl_es2_texture;
the rendering backend doesn't yet support cube maps so not decided on how to
map the samplerCube type onto this, but it might actually just be a pointer
to the first in an array of these (e.g. pointer to Px, and mandate the sequence
for a cubemap is PxNxPyNxPzNz).