glGetUniformfv(), glGetUniformiv(), and also ...
... a fair bit of rework to get the uniforms for sampler2D and samplerCube
to contain the integer index of the texture unit, as opposed to the raw
pointer to the sampler_2d. Now, internally, of course, we still want that
raw pointer to the sampler_2d and not an integer, so the translation occurs
during the load where we use the integer from the uniform to look up and
then load a pointer into the registers. This does require us to maintain
separate loading tables for the samplers: the notion of an "active texture
unit" is something specific to the frontend (GLSL/OpenGL) and not something
the execution is bothered with, consequently we need to hide that abstraction
via a few tables.

... second change is to make the GLSL runtime value for a samplerCube to
be the pointer to the first sampler_2d in an array of 6, one for each
face of the cube. To make that array explicit, rather than implicit, the
gl_es2_texture now contains an explicit texture_cube_maps_[6] array where
before there were 6 consecutive fields.

... third change is, if the integer value for a sampler is somehow bogus,
it will translate (during uniform load) to a NULL (yay), however, the
texture2D builtin would then crash on the NULL. Made it so texture2D will
ignore NULLs; the resulting texture value will then be undefined (whatever
garbage was in the destination register.)