Make GLES2/gl.h header inclusion technique compile-time switchable...
... one way (the way it works inside aex-gl), we define all functions as local
static function pointers to the aex_gl_es2_XXX function equivalent, the other
way, we expect the glXXX function to be linked.

This is so we can run and compare demo output with non-AEX-GL platforms, as
well as change our link approach in the future (e.g. if we start exporting
GLESv2 symbols without concern for any linkage namespace collisions.)