Make fragment buffer have a single triangle orientation for all fragments ..
.. right now we always rasterize all fragments for a triangle before moving to
the next. That is inefficient. What we'd like to do is to rasterize all
fragments (buffer size permitting) for all triangles in a single draw call, as
those fragments typically all have the same shader execution, blend, and
stencil characteristics. Except for the one detail that the blend and stencil
characteristics are conditional upon the orientation of the triangle.
We don't want to switch on the orientation for each fragment (that'd be
incredibly wasteful!), nor do we want to resign ourselves to per-triangle
processing (that would be efficient for what I'd expect to be very common use-
cases.)
So, we mandate that a fragment buffer will only contain the fragments of the
same single orientation, and will stall the triangle rasterization if we
encounter a mismatch.