line drawing and blending fixes ..
.. also primitive_assembly_draw_elements() has now been replaced.

The line drawing fixes are mostly in the formation of the triangulation for the
line segments.

Additionally, the blend function has been fixed; we previously read the
background (destination) values for each fragment such that we did not have to
pointer-indirect those values in the middle of blending loops. Unfortunately,
this fails if the pixel being read is a pixel that was written to in the same
batch (e.g. shared edges of the same draw call) - this popped up very visibly
in line drawing.

Line drawing now works to spec on the simple sixteen_lines demo, but has issues
with the world_is_yours demo (so more investigation needed.)