AexGL-DLL project now also builds EGL.dll and GLESv2.dll forwarding DLLs ...
... they forward into AexGL.dll and have no code. Them not having any code is
valid but something the visual studio tooling borks over; two problems being
mitigated.

First, if we pass the .def file directly to the linker, it thinks the symbols
in the def file should appear in an object file (it should not, because they're
forwarded, but the linker is not smart enough to figure this out.) The
solution is to pass it an .exp file as generated by the lib tool instead.

Second problem, if we pass an .exp file to the linker within its project, but
without any source code object files, then msbuild or visual studio (not sure)
will never build this configuration.

We work around this by performing a manual build step; unfortunately a manual
build step needs to be a single command, whereas we need to execute two
commands, consequently the commands to execute are in a batch file, and we
call that batch file instead.

All this is terribly ugly but the end result is very nice; both EGL.dll and
GLESv2.dll are small (2KB and 9KB respectively) and forward their calls to
one and the same AexGL.dll in the same directory. Just very nice and tight.