eglMakeCurrent(), eglSwapBuffers()...
... just enough EGL support to get us running in a GLFW demo (see readme.md)

Couple of issues here, first, this is not yet firm to the EGL spec, we need to
travel the configuration path and implement all the APIs.
Second, the implementation eglSwapBuffers is really quite horrendously awful.
In particular, the way we blit from the framebuffer to the screen is fairly
ghastly. This is just to get it working, a better solution would be to use
a DIBSection or DirectX surface or such, probably going with a DIBSection for
ease of use.
However I believe a DIBSection either requires, or works best with, BGRA,
whereas we write/blend into the framebuffer as RGBA, so may need to switch
the blend backend to BGRA by flag.
Third problem is the current resize behavior is wrong, I believe when the HWND
resizes, the framebuffer should automatically resize with it on the backend,
while preserving the previous viewport. Either we re-allocate the storage for
the framebuffer with the glViewport call, or we re-allocate the storage after
we do the eglSwapBuffers (though then it'll get it wrong for 1 frame.) Need
to think about the proper behavior.