Add is_indirect_ flag to reg_alloc (WIP) and variety of flag init fixes..
.. the "is_indirect_" flag intends to model the scenario where the registers
don't hold the value directly, but instead are integer registers that point
to the registers (float/bool/int/whatnot) that hold the actual values.

This is a foray into the following problem:
mat4 m[6];

m[3][2].yzwx[3]; /* <-- this is a valid lvalue */

The above is the third column (#2) of the fourth matrix (#3), with the
components swizzled, after which we select the fourth (#3) coordinate,
which after swizzling, is the X coordinate.

This expression is trivial as an rvalue, but hard as an lvalue, and, by
my reading of the spec, we need to support it. Going the full indirection
route is slow but necessary because the swizzling mucks things up, otherwise,
we could have treated vecs and mats as special types of arrays, but alas, now
it doesn't bring us closer.

Work in progress, may revert if it does not get us there.