rasterizer.c must not be built for edit & continue for x86 ..
.. the problem with edit & continue is it turns the __LINE__ preprocessor
object macro from a constant into a variable (the idea being the line# might
change when you edit the source.)
However, we rely on it as a constant as we use it as a case label.

For x64 we already did this, but x86 failed the build. x86 is not an explicit
target (as everything is designed for 64 bit registers) but it should at least
build.