rvalue fixes ...
when we do:
vec4 v = vec4(..);
v.wzyx = v.xyzw;    <--- swizzle here is what matters

the swizzle assignment gave us a problem, in particular, because we did not
reserve any rvalue registers for it, we were overwriting components w and z
before we could read from them to assign it to y and x (so you got xyyx).

Fixed by forcing exop_assign to allocate an rvalue for what it's trying to
assign regardless. Right now that's wasteful, but when we add a register
coloring stage at a later point when we start caring about performance, the
rvalues will be move-elided when not necessary.

Now.. forcing rvalues on assignments exposed a *lot* of cases where we were
writing the result of an expression evaluation into the "rvalue if available,
regular base reg otherwise". This went undetected before because we rarely
required an rvalue (and so it didn't cause any problems.)

Fixed by using the appropriate access macro; unfortunately in a lot of places.