sl_reg_move() now also handles simple scalar type conversions ..
.. so from int, to float, from float, to bool, from bool, to int, and so on.
The conversion "is what you would expect", i.e. non-zero is true, zero is false,
true converts to 1 or 1.f, false converts to 0 or 0.f.

Note that the "true" value here is 0x01, not 0xFF; may want to change that
later if we're going to be using it for masking.