function-definition-setup non-terminal destructor should unwind both scope ...
... and frame. The scope was previously unwound in the production for
function-definition, but that does not get called when in error recovery.
Did not unwind the frame to the global frame, which cause global variables
after a function definition to be attached to the last function definition
instead of the global frame. That caused functionality that examined the
global frame (e.g. uniform & attribute processing) to miss out on these
variables for processing.