fix for counterclockwise varyings mismatching their vertex ..
.. previously when we noticed we were counterclockwise we would swap two
vertices so all the Dp01, Dp12, and Dp20 would be positive when inside the
triangle and negative outside the triangle...

... unfortunately that would also require swapping out all the varying
attributes, otherwise they would unintentionally switch vertices whenever
a triangle was counterclockwise. That we did not do ..

.. new fix is to recognize that it does not matter if a pixel is on the
*positive* side of each edge of the triangle, what matters instead is that
a pixel is on the *same* side of each edge of the triangle, as it can then
only be inside the triangle ..

.. Old swapping of vertices removed, D012 still flipped for the extra bit
of precision, and the inner loop has a different mask switch that checks
if we're on the same side of each edge, rather than only of we're on the
positive side of each edge. Should still parallelize nicely, slightly
higher cost for generalism.