GL ES2 API log functionality (WIP) ..
.. idea here is that we log, in detail, what is being called, and perhaps at
a later stage do a "picture framer" where we can re-create problematic frames
for analysis outside of the original env.

.. here we dispatch on each entry to a log function. We'll put that behind a
flag later. The alternative is to call the log function and have it trap-door
into the implementation function (so you get the return value) - while the
benefit would be that it decouples the logging from the implementation (and
even allow logging of non-Aex-GL implementations), it is more costly in terms
of call overhead (because we dispatch even when not logging) and to overcome
that call overhead we would have separate logging and non-logging versions. So,
going with this "least surprising" approach for now, may switch later.