fix: if the framebuffer height is odd, you'll need to have an oversized ..
.. memory region for one additional row of pixels on any of the textures, or
renderbuffers attached to the framebuffer.

Reason for this is that we rasterize per 4 pixels (in a square) to accomplish
mip-mapping and dFdx()/dFdy() type shader operations. While the access to this
row is masked (and no bytes are ever changed), the bytes do need to be
accessible, otherwise we'll run into stray memory accesses.