Support for glReadPixels(.., GL_DEPTH_COMPONENT, ..) and a fix...
... the glReadPixels() feature allows us debug access to the zbuffer contents
while not having to pry open the interface (can just use OpenGL); is it non-
standard. The datatype size that you read vs. the datatype size of the zbuffer
are auto-scaled, so, for instance, suppose the zbuffer is 16 bits and we read
GL_UNSIGNED_INT 32 bits, then the 16 bits will be duplicated such that values
span the full 32 bit range, but without loss of precision.