ir.c/h: instruction based intermediate representation form (WIP) ..
.. kind of reluctant to go in this direction because of the big ball of
complexity that it'll create, but I think we've reached the limits of executing
from inside the AST. In particular, from a debugging perspective, having an
instruction based form as human-readable output will be huge, compared to
traversing the AST at runtime in debug sessions.
.. I think this IR form, when complete, should be able to take us all the way
to JIT binaries, that is to say, it should be able to fit machine instructions
like X64. That said, I'm not sure we need to take it there, we might want to
go from running out of the AST to running out of the IR first.

Work in progress, still needs instruction registries and linkable objects
(lob_sym is a symbol from a linkable object.)