simplify internal triangle rasterizer pixel-coverage condition ..
.. so before we had "inside all edges, or outside all edges, is inside
triangle" - this notion of "outside all edges" reflected that if the edges were
counterclockwise, then all the determinants would flip their sign, and so the
area inside the triangle would be the negative side of all edges..

.. however, we since changed the determinants when things are counterclockwise
as part of fixing the varying and Z-buffer interpolations. This has the added
benefit that the area inside the triangle is now again the positive side of all
the edges.

.. thus we revert the condition to the simpler check for positive on all edges,
this is the condition we originally had (when we swapped two vertices to get
the counterclockwise coverage, which was an incorrect technique due to varying
interpolation.)

.. the condition is now simpler, should also be faster, in the innermost part
of the loop.